+++ shims/spark300/src/main/scala/com/nvidia/spark/rapids/shims/spark300/Spark300Shims.scala
@@ -200,7 +200,7 @@ class Spark300Shims extends SparkShims {
             .ansiEnabled, conf, p, r)),
       GpuOverrides.expr[AnsiCast](
         "Convert a column of one type of data into another type",
-        new CastChecks(),
+        new CastChecks(ansiMode = Some(true)),
         (cast, conf, p, r) => new CastExprMeta[AnsiCast](cast, true, conf, p, r)),
       GpuOverrides.expr[TimeSub](
         "Subtracts interval from timestamp",
