+++ sql-plugin/src/main/scala/com/nvidia/spark/rapids/RapidsConf.scala
@@ -484,9 +484,15 @@ object RapidsConf {
     .booleanConf
     .createWithDefault(false)
 
+  val ENABLE_CAST_FLOAT_TO_DECIMAL = conf("spark.rapids.sql.castFloatToDecimal.enabled")
+    .doc("Casting from floating point types to decimal on the GPU returns results that have " +
+      "tiny difference compared to results returned from CPU.")
+    .booleanConf
+    .createWithDefault(false)
+
   val ENABLE_CAST_FLOAT_TO_STRING = conf("spark.rapids.sql.castFloatToString.enabled")
     .doc("Casting from floating point types to string on the GPU returns results that have " +
-      "a different precision than the default Java toString behavior.")
+      "a different precision than the default results of Spark.")
     .booleanConf
     .createWithDefault(false)
 
@@ -497,6 +503,12 @@ object RapidsConf {
           "documentation for more details.")
       .booleanConf
       .createWithDefault(false)
+
+  val ENABLE_CAST_LONG_TO_DECIMAL = conf("spark.rapids.sql.castLongToDecimal.enabled")
+    .doc("Long values which contain more than 18 digits can not be converted to decimal on " +
+      "the GPU, because max supported precision of decimal under GPU is 18.")
+    .booleanConf
+    .createWithDefault(false)
 
   val ENABLE_CAST_STRING_TO_FLOAT = conf("spark.rapids.sql.castStringToFloat.enabled")
     .doc("When set to true, enables casting from strings to float types (float, double) " +
@@ -1022,7 +1034,11 @@ class RapidsConf(conf: Map[String, String]) extends Logging {
 
   lazy val enableHashOptimizeSort: Boolean = get(ENABLE_HASH_OPTIMIZE_SORT)
 
+  lazy val isCastFloatToDecimalEnabled: Boolean = get(ENABLE_CAST_FLOAT_TO_DECIMAL)
+
   lazy val isCastFloatToStringEnabled: Boolean = get(ENABLE_CAST_FLOAT_TO_STRING)
+
+  lazy val isCastLongToDecimalEnabled: Boolean = get(ENABLE_CAST_LONG_TO_DECIMAL)
 
   lazy val isCastStringToTimestampEnabled: Boolean = get(ENABLE_CAST_STRING_TO_TIMESTAMP)
